refactor(selection): story_selection_panel.dart 분해 (1648→561줄)
이미 16개 private 위젯으로 잘 나뉘어 있던 story_selection_panel.dart를
3개 part 파일로 물리적 분할. 코드 변경 0건, 라이브러리 프라이빗 유지.

신설:
- lib/widgets/selection/panel_chrome.dart  (~280줄)
  · _stageAccentColor, _PanelTexturePainter, _PanelTextureLayer,
    _BottomFadeHint, _CollapsedSelectionPeek, _SelectionSheetTopBar,
    _SelectionSheetHandle, _SheetStepButton
- lib/widgets/selection/step_chip.dart  (~340줄)
  · _StepChip, _StepSummaryPill, _CompactSegmentedToggle,
    _CompactSegmentedToggleButton, _PrimaryActionButton, _EmptyStepMessage
- lib/widgets/selection/selection_cards.dart  (~465줄)
  · _CardShell, _EraCompactCard, _PersonCompactCard, _StoryCompactCard,
    _IndexBadge, _PersonDots, _PortraitAvatar, _AvatarFallback

부모 파일:
- part 'selection/panel_chrome.dart' 등 디렉티브 추가
- 본체는 StorySelectionPanel + State + 시간 포맷 헬퍼만 보유

flutter analyze: 0 issues
flutter test: 80 tests pass
